Use "JSON for Modern C++" for returning the JSON responses
Also, ditch the error handling with returning boolean and setting an
error message:

  bool Id(result, std::string* err);

Provided that the libraries that are used (std::* and JSON for Modern
C++) throw exceptions, there is no need to swim against the current and
catch their exceptions inside the IPFS methods. Use:

  void Id(result);

instead and throw an exception on failure.