Change the Coveralls repo token
I had to delete the repo from Coveralls in order to remove a mistakenly
added test/utils.h. Even adding it in "coveralls --exclude ..." does not
remove it from the list of the files on Coveralls, even worse, it starts
showing it at 0% coverage because no data is sent for it anymore.